+++ tests/run_tests.py
@@ -25,6 +25,7 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 import argparse
 import datetime
+import difflib
 import json
 import logging
 import os
@@ -168,6 +169,7 @@ class TestRunner(object):
         console_handler.setFormatter(formatter)
 
         self.logger.addHandler(console_handler)
+        #self.logger.addHandler(logging.StreamHandler())
 
     def run(self):
         return self.__run()
@@ -227,12 +229,22 @@ class TestRunner(object):
 
         testcase_inputpath = os.path.join(self.input_path, testcase_inputpath)
 
+        self.logger.info('Invoking test case \"{}\"'.format(testcase_name))
+
         if not os.path.exists(testcase_inputpath) or not os.path.isfile(testcase_inputpath):
             self.__log_error('Test case \"{}\" does not have valid input'.format(testcase_name))
             return self.StatusCode.ERROR
 
-        cmd = '{executable} --errors --output ./ {input_path}'.format(
+        output_dir_name = 'output'
+
+        output_dir_path = os.path.join(os.getcwd(), output_dir_name)
+
+        if not os.path.exists(output_dir_path):
+           os.mkdir(output_dir_path)
+
+        cmd = '{executable} --errors --no-annotation-comments --output {output_dir_path} {input_path}'.format(
             executable=self.executable_invoke_name,
+            output_dir_path=output_dir_path,
             input_path=testcase_inputpath)
 
         completed_process = subprocess.run(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
@@ -252,6 +264,10 @@ class TestRunner(object):
             ))
             err = self.StatusCode.ERROR
 
+
+        if err == self.StatusCode.SUCCESS and return_code == 0:
+            err = self.__run_snapshot_test_for_testcase(testcase, output_dir_path)
+
         if err == self.StatusCode.SUCCESS:
             self.__log_success('[{}] - {}'.format(testcase_name, self.StatusCode.tostring(err)))
         else:
@@ -265,6 +281,62 @@ class TestRunner(object):
                     testcase_inputpath=testcase_inputpath))
 
         return err
+
+    def __run_snapshot_test_for_testcase(self, testcase, output_dir_path):
+        testcase_snapshot_filename_without_ext = testcase['output_file_name_without_ext']
+
+        integration_test_dir = os.path.dirname(os.path.realpath(__file__))
+
+        snapshot_file_dir = os.path.join(integration_test_dir, 'snapshots')
+
+        snapshot_header_file_path = os.path.join(snapshot_file_dir, testcase_snapshot_filename_without_ext + '.h')
+
+        if not os.path.exists(snapshot_header_file_path):
+            self.__log_error('Cannot find {}'.format(snapshot_header_file_path))
+            return self.StatusCode.ERROR
+
+        snapshot_cpp_file_path = os.path.join(snapshot_file_dir, testcase_snapshot_filename_without_ext + '.cpp')
+
+        if not os.path.exists(snapshot_cpp_file_path):
+            self.__log_error('Cannot find {}'.format(snapshot_cpp_file_path))
+            return self.StatusCode.ERROR
+
+        output_header_file_path = os.path.join(output_dir_path, testcase_snapshot_filename_without_ext + '.h')
+
+        if not os.path.exists(output_header_file_path):
+            self.__log_error('Cannot find {}'.format(output_header_file_path))
+            return self.StatusCode.ERROR
+        
+        output_cpp_file_path = os.path.join(output_dir_path, testcase_snapshot_filename_without_ext + '.cpp')
+
+        if not os.path.exists(output_cpp_file_path):
+            self.__log_error('Cannot find {}'.format(output_cpp_file_path))
+            return self.StatusCode.ERROR
+
+        err = self.__diff_two_files(snapshot_header_file_path, output_header_file_path)
+
+        if err != self.StatusCode.SUCCESS:
+            return err
+
+        err = self.__diff_two_files(snapshot_cpp_file_path, output_cpp_file_path)
+
+        return err
+
+    def __diff_two_files(self, file1, file2):
+        text1 = open(file1).readlines()
+        text2 = open(file2).readlines()
+
+        has_diff = False
+        shown_log = False
+
+        for line in difflib.unified_diff(text1, text2):
+            if not shown_log:
+                self.__log_info('Diffing between \"{}\" and \"{}\"'.format(file1, file2))
+                shown_log = True
+            self.__log_info(line)
+            has_diff = True
+
+        return self.StatusCode.SUCCESS if has_diff is False else self.StatusCode.FAILURE
 
     def __formatted_errors(self, input_path, expected_errors):
         if not expected_errors:
@@ -303,6 +375,13 @@ class TestRunner(object):
     def __log_msg(self, msg, color=None):
         if color:
             self.console_logger.info('{}{}{}'.format(color, msg, Colors.ENDC))
+        else:
+            self.console_logger.info(msg)
+        self.logger.info(msg)
+
+    def __log_info(self, msg):
+        if self.opts.colors:
+            self.console_logger.info(msg)
         else:
             self.console_logger.info(msg)
         self.logger.info(msg)